+++ lib/ravelry/errors/configuration.rb
@@ -1,0 +1,5 @@
+module Ravelry
+  module Errors
+    class Configuration < StandardError; end
+  end
+end