+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.1.0
+
+- Return `self` on both `Ravelry::Pattern.get` and `Ravelry::Pattern.permalink_get` (previously returned `data`)
+- Add `callback_url` configuration option for OAuth
+- Create `Ravelry::Error::Configuration` class for configuration errors
+- Add more `Ravelry::Pattern` documentation
+
 ## 0.0.9
 
 - Add `json` instance method to `Ravelry::Pattern`